Reviewed Instructions and code in the javascript file. Created another js file for my data.
My data for this project will most likely be the books that I have read this year. And will probably expand to books that I own. I've been meaning to create a catalog of books I own anyway, I was thinking just on google sheets but this is a great opportunity for me to try it on a website!