Adding data form, URLs to array format, js function to console log form info
Created a little form in preparation of adding data more easily. Saved URLs into an array so that it is easier to go through the list of urls. My function to add cards passed the testing phase, so I now know how to access the information inside the text boxes and can console log it.